refactor(generic): move ora.ora to an ora helper for ease of submodule use
ISSUES CLOSED: #100
