+++ src/util/terminate.js
@@ -1,8 +1,4 @@
 import colors from 'colors';
-import debug from 'debug';
-import ora from 'ora';
-
-const d = debug('electron-forge:lifecycle');
 
 process.on('unhandledRejection', (err) => {
   if (err && err.message && err.stack) {
@@ -27,30 +23,3 @@ process.on('uncaughtException', (err) => {
   }
   process.exit(1);
 });
-
-if (process.env.DEBUG && process.env.DEBUG.includes('electron-forge')) {
-  console.warn('WARNING: DEBUG environment variable detected.  Progress indicators will be sent over electron-forge:lifecycle'.red);
-  ora.ora = (name) => {
-    const fake = {
-      start: () => {
-        d('Process Started:', name);
-        return fake;
-      },
-      fail: () => {
-        d(`Process Failed: ${name}`.red);
-        return fake;
-      },
-      succeed: () => {
-        d('Process Succeeded:', name);
-        return fake;
-      },
-      stop: () => {
-        d('Process Stopped:', name);
-        return fake;
-      },
-    };
-    return fake;
-  };
-} else {
-  ora.ora = ora;
-}
